devtools/shared/webconsole/test/chrome/test_jsterm_cd_iframe.html

<!DOCTYPE HTML>
<html lang="en">
<head>
  <meta charset="utf8">
  <title>Test for the cd() function</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="common.js"></script>
  <!-- Any copyright is dedicated to the Public Domain.
     - http://creativecommons.org/publicdomain/zero/1.0/ -->
</head>
<body>
<p>Test for the cd() function</p>

<iframe id="content-iframe" src="http://example.com/chrome/devtools/shared/webconsole/test/chrome/sandboxed_iframe.html"></iframe>

<script class="testbody" type="text/javascript">
"use strict";

SimpleTest.waitForExplicitFinish();

let gState;

function startTest()
{
  removeEventListener("load", startTest);

  attachConsoleToTab([], onAttach);
}

function onAttach(aState, aResponse)
{
  top.foobarObject = Object.create(null);
  top.foobarObject.bug609872 = "parent";

  window.foobarObject = Object.create(null);
  window.foobarObject.bug609872 = "child";

  gState = aState;

  const tests = [doCheckParent, doCdIframe, doCheckIframe,
      doCdContentIframe,
      doCdSandboxedIframe, doCheckSandboxedIframe,
      doCdParent,
      doCdParent,
      doCheckParent2];
  runTests(tests, testEnd);
}

function doCheckParent()
{
  info("check parent window");
  gState.webConsoleFront.evaluateJSAsync("window.foobarObject.bug609872")
    .then(onFooObjectFromParent);
}

function onFooObjectFromParent(aResponse)
{
  checkObject(aResponse, {
    input: "window.foobarObject.bug609872",
    result: "parent",
  });

  ok(!aResponse.exception, "no eval exception");
  ok(!aResponse.helperResult, "no helper result");

  nextTest();
}

function doCdIframe()
{
  info("test cd('iframe')");
  gState.webConsoleFront.evaluateJSAsync("cd('iframe')").then(onCdIframe);
}

function onCdIframe(aResponse)
{
  checkObject(aResponse, {
    input: "cd('iframe')",
    result: { type: "undefined" },
    helperResult: { type: "cd" },
  });

  ok(!aResponse.exception, "no eval exception");

  nextTest();
}

function doCheckIframe()
{
  info("check foobarObject from the iframe");
  gState.webConsoleFront.evaluateJSAsync("window.foobarObject.bug609872")
    .then(onFooObjectFromIframe);
}

function onFooObjectFromIframe(aResponse)
{
  checkObject(aResponse, {
    input: "window.foobarObject.bug609872",
    result: "child",
  });

  ok(!aResponse.exception, "no js eval exception");
  ok(!aResponse.helperResult, "no helper result");

  nextTest();
}

function doCdContentIframe()
{
  info("test cd('#content-iframe')");
  gState.webConsoleFront.evaluateJSAsync("cd('#content-iframe')").then(onCdContentIframe);
}

function onCdContentIframe(aResponse)
{
  checkObject(aResponse, {
    input: "cd('#content-iframe')",
    result: { type: "undefined" },
    helperResult: { type: "cd" },
  });

  ok(!aResponse.exception, "no eval exception");

  nextTest();
}
function doCdSandboxedIframe()
{
  // Don't use string to ensure we don't get security exception
  // when passing a content window reference.
  const cmd = "cd(document.getElementById('sandboxed-iframe').contentWindow)";
  info("test " + cmd);
  gState.webConsoleFront.evaluateJSAsync(cmd).then(onCdSandboxedIframe.bind(null, cmd));
}

function onCdSandboxedIframe(cmd, aResponse)
{
  checkObject(aResponse, {
    input: cmd,
    result: { type: "undefined" },
    helperResult: { type: "cd" },
  });

  ok(!aResponse.exception, "no eval exception");

  nextTest();
}

function doCheckSandboxedIframe()
{
  info("check foobarObject from the sandboxed iframe");
  gState.webConsoleFront.evaluateJSAsync("window.foobarObject.bug1051224")
    .then(onFooObjectFromSandboxedIframe);
}

function onFooObjectFromSandboxedIframe(aResponse)
{
  checkObject(aResponse, {
    input: "window.foobarObject.bug1051224",
    result: "sandboxed",
  });

  ok(!aResponse.exception, "no js eval exception");
  ok(!aResponse.helperResult, "no helper result");

  nextTest();
}

function doCdParent()
{
  info("test cd() back to parent");
  gState.webConsoleFront.evaluateJSAsync("cd()").then(onCdParent);
}

function onCdParent(aResponse)
{
  checkObject(aResponse, {
    input: "cd()",
    result: { type: "undefined" },
    helperResult: { type: "cd" },
  });

  ok(!aResponse.exception, "no eval exception");

  nextTest();
}

function doCheckParent2()
{
  gState.webConsoleFront.evaluateJSAsync("window.foobarObject.bug609872")
    .then(onFooObjectFromParent2);
}

function onFooObjectFromParent2(aResponse)
{
  checkObject(aResponse, {
    input: "window.foobarObject.bug609872",
    result: "parent",
  });

  ok(!aResponse.exception, "no eval exception");
  ok(!aResponse.helperResult, "no helper result");

  nextTest();
}

function testEnd()
{
  closeDebugger(gState, function() {
    gState = null;
    SimpleTest.finish();
  });
}

addEventListener("load", startTest);
</script>
</body>
</html>
